Django-signals_orm-0x04/messaging_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==21.2.0
Automat==20.2.0
autopep8==2.3.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cached-property==2.0.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.0.3
cloud-init==25.1.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.5.0
genanki==0.13.1
gTTS==2.5.4
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.31.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.9.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.78.1
packaging==25.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.14.0
pycurl==7.44.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dotenv==1.1.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
systemd-python==234
tomli==2.2.1
tqdm==4.67.1
Twisted==22.1.0
typing-inspection==0.4.0
typing_extensions==4.13.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.0
wadllib==1.3.6
watchdog==6.0.0
zipp==1.0.0
zope.interface==5.4.0
